refactor solar code
