+++ Sources/App/Helper/Solar/SolarPositionAlgorithm.swift
